requirements.txt
gradio
langchain
langchain-core
langchain-groq
python-dotenv